ACP issue template: Fixed dead link
There was a link to a section about ACPs in the std-dev-guide. I rewrote the
issue template a bit as the exact section does not exist anymore. The issue
template now links to the chapter about the feature lifecycle instead (that
chapter also contains information about FCPs).

Also, fixed a minor grammatical error.